refactor(step functions): move shared declarations to a seperate folder
