+++ microservices/site-launch/lambda-functions/primary-domain-validation/index.ts
@@ -10,32 +10,16 @@ import type {
   GetDomainAssociationCommandOutput,
 } from "@aws-sdk/client-amplify"
 
-
-import logger from "../../shared/logger"
 import {
-  SITE_LAUNCH_LAMBDA_TYPE,
+  MessageBody,
   SITE_LAUNCH_LAMBDA_STATUS,
-} from "../../shared/types" // TODO: Change to aliased imports
+  SITE_LAUNCH_LAMBDA_TYPE,
+} from "../../shared/types"
 
-export interface MessageBody {
-  repoName: string
+interface PrimaryDomainValidationLambdaParams {
   appId: string
   primaryDomainSource: string
   primaryDomainTarget: string
-  domainValidationSource: string
-  domainValidationTarget: string
-  requestorEmail: string
-  agencyEmail: string
-  githubRedirectionUrl?: string
-  redirectionDomain?: [
-    {
-      source: string
-      target: string
-      type: string
-    }
-  ]
-  success?: boolean
-  siteLaunchError?: string
 }
 
 interface PrimaryDomainValidationLambdaResponse {
