+++ microservices/site-launch/lambda-functions/redirection-domain-validation/index.ts
@@ -3,28 +3,8 @@
 import { Octokit } from "@octokit/rest"
 
 import logger from "../../shared/logger"
+import { MessageBody } from "../../shared/types"
 
-
-export interface MessageBody {
-  repoName: string
-  appId: string
-  primaryDomainSource: string
-  primaryDomainTarget: string
-  domainValidationSource: string
-  domainValidationTarget: string
-  requestorEmail: string
-  agencyEmail: string
-  githubRedirectionUrl?: string
-  redirectionDomain?: [
-    {
-      source: string
-      target: string
-      type: string
-    }
-  ]
-  success?: boolean
-  siteLaunchError?: string
-}
 export const redirectionDomainValidation = async (
   event: Pick<
     MessageBody,
