+++ microservices/site-launch/lambda-functions/step-functions-trigger/index.ts
@@ -3,27 +3,8 @@
 import AWS, { Lambda, StepFunctions } from "aws-sdk"
 
 import logger from "../../shared/logger"
+import { MessageBody } from "../../shared/types"
 
-export interface MessageBody {
-  repoName: string
-  appId: string
-  primaryDomainSource: string
-  primaryDomainTarget: string
-  domainValidationSource: string
-  domainValidationTarget: string
-  requestorEmail: string
-  agencyEmail: string
-  githubRedirectionUrl?: string
-  redirectionDomain?: [
-    {
-      source: string
-      target: string
-      type: string
-    }
-  ]
-  success?: boolean
-  siteLaunchError?: string
-}
 export const stepFunctionsTrigger = async (event: MessageBody) => {
   const { AWS_REGION, AWS_ACCOUNT_NUMBER, STATE_MACHINE_NAME } = process.env
   try {
